Refactor layouts file
