+++ src/app/classes/class/[slug]/student-schema.ts
